# checkout/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/cart:
        post:
            tags:
                - CartService
            operationId: CartService_AddItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.order.v1.AddItemReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.order.v1.AddItemResp'
    /v1/cart/{name}:
        get:
            tags:
                - CartService
            operationId: CartService_GetCart
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                - name: owner
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.order.v1.GetCartResp'
        delete:
            tags:
                - CartService
            operationId: CartService_EmptyCart
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                - name: owner
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.order.v1.EmptyCartResp'
...
    /v1/credit_cards:
        post:
            tags:
                - CreditCardsService
            operationId: CreditCardsService_CreateCreditCard
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.credit_cards.v1.CreditCards'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.credit_cards.v1.CardsReply'
        patch:
            tags:
                - CreditCardsService
            operationId: CreditCardsService_UpdateCreditCard
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.credit_cards.v1.CreditCards'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.credit_cards.v1.CardsReply'
    /v1/credit_cards/all:
        get:
            tags:
                - CreditCardsService
            operationId: CreditCardsService_ListCreditCards
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.credit_cards.v1.ListCreditCardsReply'
    /v1/credit_cards/{credit_card_number}:
        get:
            tags:
                - CreditCardsService
            operationId: CreditCardsService_GetCreditCard
            parameters:
                - name: credit_card_number
                  in: path
                  required: true
                  schema:
                    type: string
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.credit_cards.v1.GetCreditCardsReply'
    /v1/credit_cards/{id}:
        delete:
            tags:
                - CreditCardsService
            operationId: CreditCardsService_DeleteCreditCard
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.credit_cards.v1.CardsReply'
...
components:
    schemas:
...
        api.credit_cards.v1.CardsReply:
            type: object
            properties:
                message:
                    type: string
                code:
                    type: integer
                    format: int32
        api.credit_cards.v1.CreditCards:
            type: object
            properties:
                credit_card_number:
                    type: string
                credit_card_cvv:
                    type: string
                credit_card_expiration_year:
                    type: string
                credit_card_expiration_month:
                    type: string
                owner:
                    type: string
                name:
                    type: string
                id:
                    type: integer
                    format: uint32
        api.credit_cards.v1.GetCreditCardsReply:
            type: object
            properties:
                credit_cards:
                    $ref: '#/components/schemas/api.credit_cards.v1.CreditCards'
        api.credit_cards.v1.ListCreditCardsReply:
            type: object
            properties:
                credit_cards:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.credit_cards.v1.CreditCards'
        api.order.v1.AddItemReq:
            type: object
            properties:
                name:
                    type: string
                owner:
                    type: string
                item:
                    $ref: '#/components/schemas/api.order.v1.CartItem'
        api.order.v1.AddItemResp:
            type: object
            properties: {}
        api.order.v1.Cart:
            type: object
            properties:
                name:
                    type: string
                owner:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.order.v1.CartItem'
        api.order.v1.CartItem:
            type: object
            properties:
                product_id:
                    type: integer
                    format: uint32
                quantity:
                    type: integer
                    format: uint32
        api.order.v1.EmptyCartResp:
            type: object
            properties: {}
        api.order.v1.GetCartResp:
            type: object
            properties:
                cart:
                    $ref: '#/components/schemas/api.order.v1.Cart'
...